requirements.txt
fastapi
uvicorn
jinja2
openai
sentence-transformers
faiss-cpu
aiofiles
transformers
numpy
python-multipart